instagram-bot/requirements.txt
aiogram>=3.0
python-dotenv
yt-dlp
aiohttp